Functional rough prototype
Connect a tag with the info written to it, and it will scan for the bluetooth address, connect, read the characteristic data, and send the updated data to the server.

However, I'm almost positive multiple connections is going to break this. Need to get a second coaster device to test. Also a lot of code that needs to be cleaned up